requirements.txt
tensorflow==2.10.1
scikit-learn==1.6.1
matplotlib==3.9.0
numpy==1.24.3
pillow==11.3.0
streamlit==1.22.0
opencv-python==4.11.0.86
